Fixes for the issue.
Due to boundary condition - when no of records was exactly in multiple
of batch-size, instead of breaking out of the loop, we were querying
cassandra one more time. Which was returning null - if no records found,
same situation was not handled.
Now handling null return value as well as fixing the boundray condition.